[MODFISTO-101] - Return helpful and clear error code, when Group name already exist
