+++ src/test/java/org/folio/rest/impl/TestBase.java
@@ -40,9 +40,9 @@ import io.vertx.core.logging.LoggerFactory;
 public abstract class TestBase {
   protected final Logger logger = LoggerFactory.getLogger(this.getClass());
 
-  public static final String TENANT_NAME = "diku";
-  public static final String NON_EXISTED_ID = "bad500aa-aaaa-500a-aaaa-aaaaaaaaaaaa";
-  public static final Header TENANT_HEADER = new Header(OKAPI_HEADER_TENANT, TENANT_NAME);
+  static final String TENANT_NAME = "diku";
+  static final String NON_EXISTED_ID = "bad500aa-aaaa-500a-aaaa-aaaaaaaaaaaa";
+  static final Header TENANT_HEADER = new Header(OKAPI_HEADER_TENANT, TENANT_NAME);
 
   private static boolean invokeStorageTestSuiteAfter = false;
